Extract files into commons for tests
